refactor(semantic): unify zsh plugin framework contracts
