+++ crates/shuck-semantic/src/source_closure/plugin_managers/mod.rs
@@ -10,8 +10,10 @@ mod generic_zsh_runtime;
 mod oh_my_zsh;
 mod prezto;
 mod zdot;
+mod zinit;
 
 use super::*;
+use crate::ZshPluginFramework;
 
 pub(super) use oh_my_zsh::{dedup_plugin_requests, sorted_dependency_paths};
 
@@ -32,20 +34,18 @@ pub(super) struct DeferredPluginRuntimeContext<'a> {
     pub(super) synthetic_reads: &'a [SyntheticRead],
 }
 
-trait ZshPluginManager {
+trait ZshPluginManager: ZshPluginFramework {
     fn is_active(&self, context: &PluginManagerContext<'_>) -> bool {
         context.model.shell_profile().dialect == ParseShellDialect::Zsh
-    }
-
-    fn is_active_for_deferred(&self, context: &DeferredPluginRuntimeContext<'_>) -> bool {
-        context.semantic.shell_profile().dialect == ParseShellDialect::Zsh
     }
 
     fn collect_plugin_requests(&self, context: &PluginManagerContext<'_>) -> Vec<PluginRequest> {
         let _ = context;
         Vec::new()
     }
+}
 
+trait ZshDeferredRuntimeManager {
     fn collect_deferred_required_reads(
         &self,
         context: &DeferredPluginRuntimeContext<'_>,
@@ -53,6 +53,40 @@ trait ZshPluginManager {
         let _ = context;
         Vec::new()
     }
+}
+
+static OH_MY_ZSH_PLUGIN_MANAGER: oh_my_zsh::OhMyZshPluginManager = oh_my_zsh::OhMyZshPluginManager;
+static PREZTO_PLUGIN_MANAGER: prezto::PreztoPluginManager = prezto::PreztoPluginManager;
+static ZDOT_PLUGIN_MANAGER: zdot::ZdotPluginManager = zdot::ZdotPluginManager;
+static ZINIT_PLUGIN_MANAGER: zinit::ZinitPluginManager = zinit::ZinitPluginManager;
+
+static ZSH_PLUGIN_MANAGERS: [&dyn ZshPluginManager; 4] = [
+    &OH_MY_ZSH_PLUGIN_MANAGER,
+    &PREZTO_PLUGIN_MANAGER,
+    &ZDOT_PLUGIN_MANAGER,
+    &ZINIT_PLUGIN_MANAGER,
+];
+
+static ZSH_PLUGIN_FRAMEWORKS: [&dyn ZshPluginFramework; 4] = [
+    &OH_MY_ZSH_PLUGIN_MANAGER,
+    &PREZTO_PLUGIN_MANAGER,
+    &ZDOT_PLUGIN_MANAGER,
+    &ZINIT_PLUGIN_MANAGER,
+];
+
+/// Returns all built-in zsh plugin framework implementations.
+pub fn zsh_plugin_frameworks() -> &'static [&'static dyn ZshPluginFramework] {
+    &ZSH_PLUGIN_FRAMEWORKS
+}
+
+/// Returns the built-in implementation for a framework, when Shuck knows one.
+pub fn layout_for_plugin_framework(
+    framework: &PluginFramework,
+) -> Option<&'static dyn ZshPluginFramework> {
+    zsh_plugin_frameworks()
+        .iter()
+        .copied()
+        .find(|layout| &layout.framework() == framework)
 }
 
 pub(super) fn collect_plugin_requests(
@@ -73,15 +107,8 @@ pub(super) fn collect_plugin_requests(
         source_path,
         plugin_resolver,
     };
-    let managers: [&dyn ZshPluginManager; 4] = [
-        &oh_my_zsh::OhMyZshPluginManager,
-        &prezto::PreztoPluginManager,
-        &zdot::ZdotPluginManager,
-        &generic_zsh_runtime::GenericZshRuntimeManager,
-    ];
-
     let mut requests = Vec::new();
-    for manager in managers {
+    for manager in ZSH_PLUGIN_MANAGERS {
         if manager.is_active(&context) {
             requests.extend(manager.collect_plugin_requests(&context));
         }
@@ -109,16 +136,12 @@ pub(super) fn deferred_zsh_entrypoint_required_reads(
         scope,
         synthetic_reads,
     };
-    let managers: [&dyn ZshPluginManager; 2] = [
-        &oh_my_zsh::OhMyZshPluginManager,
-        &generic_zsh_runtime::GenericZshRuntimeManager,
-    ];
+    let managers: [&dyn ZshDeferredRuntimeManager; 1] =
+        [&generic_zsh_runtime::GenericZshRuntimeManager];
 
     let mut reads = Vec::new();
     for manager in managers {
-        if manager.is_active_for_deferred(&context) {
-            reads.extend(manager.collect_deferred_required_reads(&context));
-        }
+        reads.extend(manager.collect_deferred_required_reads(&context));
     }
     reads.sort_by(|left, right| left.as_str().cmp(right.as_str()));
     reads.dedup();
@@ -143,4 +166,24 @@ fn static_plugin_names<'a>(names: impl Iterator<Item = &'a str>) -> Vec<String> 
         plugins.push(name.to_owned());
     }
     plugins
+}
+
+fn suffix_after_last_marker<const N: usize>(path: &str, markers: [&str; N]) -> Option<PathBuf> {
+    for marker in markers {
+        if let Some(index) = path.rfind(marker) {
+            let suffix = &path[index + 1..];
+            if !suffix.is_empty() {
+                return Some(PathBuf::from(suffix));
+            }
+        }
+    }
+    None
+}
+
+fn path_text_starts_with_path(path: &str, root: &Path) -> bool {
+    let root_text = root.to_string_lossy().replace('\\', "/");
+    path == root_text
+        || path
+            .strip_prefix(&root_text)
+            .is_some_and(|tail| tail.starts_with('/'))
 }
